Added Ruby test infrastructure and Console tests in Ruby Fixed issues identified by the new tests:   - Improper formatting of object-id in get-query   - Remote (non-broker-resident) agents not visible to the console   - object.update() and object.merge() not implemented
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@817312 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qmf/ObjectIdImpl.cpp
@@ -115,7 +115,7 @@ std::string ObjectIdImpl::asString() const
 {
     stringstream val;
 
-    val << getFlags() << "-" << getSequence() << "-" << getBrokerBank() << "-" <<
+    val << (int) getFlags() << "-" << getSequence() << "-" << getBrokerBank() << "-" <<
         getAgentBank() << "-" << getObjectNum();
     return val.str();
 }
